+++ nucypher/cli/commands/ursula.py
@@ -31,7 +31,7 @@ from nucypher.cli.actions.config import (
     get_provider_process,
     handle_missing_configuration_file
 )
-from nucypher.cli.actions.network import determine_external_ip_address
+from nucypher.utilities.networking import determine_external_ip_address
 from nucypher.cli.actions.select import select_client_account, select_config_file, select_network
 from nucypher.cli.utils import make_cli_character, setup_emitter
 from nucypher.cli.commands.deploy import option_gas_strategy
